Add Fcf.Class.Bifunctor
