+++ first-class-families.cabal
@@ -33,6 +33,7 @@ library
     Fcf.Data.Nat
     Fcf.Data.Symbol
     Fcf.Classes
+    Fcf.Class.Bifunctor
     Fcf.Class.Functor
     Fcf.Utils
   build-depends:
