+++ src/Fcf/Classes.hs
@@ -9,28 +9,15 @@
 module Fcf.Classes
   ( Map
   , Bimap
+  , First
+  , Second
   , Compare
   ) where
 
 
 import Fcf.Core
 import Fcf.Class.Functor
-
-
--- $
--- >>> import Fcf.Core
-
-
--- | Type-level 'Data.Bifunctor.bimap'.
-data Bimap :: (a -> Exp a') -> (b -> Exp b') -> f a b -> Exp (f a' b')
-
--- (,)
-type instance Eval (Bimap f g '(x, y)) = '(Eval (f x), Eval (g y))
-
--- Either
-type instance Eval (Bimap f g ('Left  x)) = 'Left  (Eval (f x))
-type instance Eval (Bimap f g ('Right y)) = 'Right (Eval (g y))
-
+import Fcf.Class.Bifunctor
 
 -- | Type-level 'Compare' for totally ordered data types.
 --
